Fix VP dragon portals
Add check for same zone to first if floor_port check; if not the same zone, use zoneid instead of zone->GetZoneID.

This was previously relying on the "double zone" behavior to function and was broken by commit e71744e.

Minor refactor to remove bool is_same_zone as it was redundant with zoneid == zone->GetZoneID() checks.